Mejorando banner

+++ farmacia-back/src/main/java/co/edu/uniandes/csw/farmacia/entities/FacturaEntity.java
@@ -19,7 +19,7 @@ import uk.co.jemos.podam.common.PodamExclude;
  */
 @Entity
 public class FacturaEntity extends BaseEntity implements Serializable{
-    private int totalFactura ;
+    private Long totalFactura ;
     
     @Temporal(javax.persistence.TemporalType.DATE)
     private Date fecha;
@@ -38,11 +38,11 @@ public class FacturaEntity extends BaseEntity implements Serializable{
         this.clientes = clientes;
     }
     
-    public int getTotalFactura() {
+    public Long getTotalFactura() {
         return totalFactura;
     }
 
-    public void setTotalFactura(int totalFactura) {
+    public void setTotalFactura(long totalFactura) {
         this.totalFactura = totalFactura;
     }
     public Date getFecha() {
