+++ farmacia-web/src/main/java/co/edu/uniandes/csw/farmacia/dtos/FacturaDTO.java
@@ -6,7 +6,6 @@
 package co.edu.uniandes.csw.farmacia.dtos;
 
 import co.edu.uniandes.csw.farmacia.entities.FacturaEntity;
-import java.text.SimpleDateFormat;
 import java.util.Date;
 
 /**
@@ -17,7 +16,7 @@ public class FacturaDTO {
    
     private Long id;
     private Date fecha;
-    private int totalfactura;
+    private Long totalfactura;
 
     public FacturaDTO(){}
     
@@ -52,11 +51,11 @@ public class FacturaDTO {
         this.fecha = fecha;
     }
 
-    public int getTotalfactura() {
+    public Long getTotalfactura() {
         return totalfactura;
     }
 
-    public void setTotalfactura(int totalfactura) {
+    public void setTotalfactura(Long totalfactura) {
         this.totalfactura = totalfactura;
     }
